add other item code to app
* refactor views
* add example fixtures
* consistent terminology for questionnaires
