--- example/fixtures/initial_data.yaml
@@ -1,127 +1,0 @@
-- fields: {admin_access_only: false, html: '', name: example, parse_html: false, redirect_url: /}
-  model: questionnaire.questionnaire
-  pk: 1
-- fields: {checks: '', heading: Page 1, parse_html: true, questionnaire: 1, sortid: 1,
-    text_de: <h2> Biervorlieben</h2>, text_en: <h2>Beer Preferences</h2>}
-  model: questionnaire.questionset
-  pk: 1
-- fields: {checks: '', heading: Page 2, parse_html: false, questionnaire: 1, sortid: 2,
-    text_de: h1. Python Web Frameworks, text_en: h1. Python Web Frameworks}
-  model: questionnaire.questionset
-  pk: 2
-- fields: {checks: '', heading: Finished!, parse_html: true, questionnaire: 1, sortid: 99,
-    text_de: "<h2> Vielen Dank </h2>\r\n \r\n Wir hoffen, dass Sie uns in Zukunft
-      wieder besuchen!", text_en: "<h2>Thank you! </h2>\r\n \r\n We hope that you
-      visit us again in the future!"}
-  model: questionnaire.questionset
-  pk: 3
-- fields: {checks: '', extra_de: '', extra_en: '', footer_de: null, footer_en: '',
-    number: '1', parse_html: false, questionset: 1, sort_id: null, text_de: 'Trinken
-      Sie Bier?', text_en: 'Do you drink beer?', type: choice-yesno}
-  model: questionnaire.question
-  pk: 1
-- fields: {checks: 'requiredif="1,yes"', extra_de: '', extra_en: '', footer_de: null,
-    footer_en: '', number: '2', parse_html: false, questionset: 1, sort_id: null,
-    text_de: 'Was Art von Bier trinken Sie am meisten?', text_en: 'What type of beer
-      do you drink predominantly?', type: choice-freeform}
-  model: questionnaire.question
-  pk: 2
-- fields: {checks: '', extra_de: '', extra_en: '', footer_de: null, footer_en: '',
-    number: '3', parse_html: false, questionset: 1, sort_id: null, text_de: 'Welche
-      von dieser Bieren haben Sie probiert?', text_en: 'Which of these brands of beer
-      have you tried?', type: choice-multiple}
-  model: questionnaire.question
-  pk: 3
-- fields: {checks: '', extra_de: '', extra_en: '', footer_de: null, footer_en: '',
-    number: '4', parse_html: false, questionset: 2, sort_id: null, text_de: 'Which
-      Python Web Frameworks have you tried?', text_en: 'Which Python Web Frameworks
-      have you tried?', type: choice-multiple-freeform}
-  model: questionnaire.question
-  pk: 4
-- fields: {checks: '', extra_de: '', extra_en: '', footer_de: null, footer_en: '',
-    number: '5', parse_html: false, questionset: 2, sort_id: null, text_de: "Welche
-      m\xF6gen Sie am liebsten?", text_en: 'Which do you like the most?', type: choice-freeform}
-  model: questionnaire.question
-  pk: 5
-- fields: {question: 2, sortid: 1, tags: '', text_de: Pils, text_en: Pils/Bitter,
-    value: pils}
-  model: questionnaire.choice
-  pk: 1
-- fields: {question: 2, sortid: 2, tags: '', text_de: Helles, text_en: Lager, value: helles}
-  model: questionnaire.choice
-  pk: 2
-- fields: {question: 2, sortid: 3, tags: '', text_de: Weizen, text_en: Wheat-beer,
-    value: weizen}
-  model: questionnaire.choice
-  pk: 3
-- fields: {question: 3, sortid: 1, tags: '', text_de: Heineken, text_en: Heineken,
-    value: heineken}
-  model: questionnaire.choice
-  pk: 4
-- fields: {question: 3, sortid: 2, tags: '', text_de: Becks, text_en: Becks, value: becks}
-  model: questionnaire.choice
-  pk: 5
-- fields: {question: 3, sortid: 3, tags: '', text_de: "L\xF6wenbr\xE4u", text_en: "L\xF6wenbr\xE4u",
-    value: lowenbrau}
-  model: questionnaire.choice
-  pk: 6
-- fields: {question: 2, sortid: 4, tags: '', text_de: "Altbier/D\xFCssel", text_en: Altbier,
-    value: altbier}
-  model: questionnaire.choice
-  pk: 7
-- fields: {question: 2, sortid: 5, tags: '', text_de: "K\xF6lsch", text_en: "K\xF6lsch",
-    value: koelsch}
-  model: questionnaire.choice
-  pk: 8
-- fields: {question: 4, sortid: 1, tags: '', text_de: Django, text_en: Django, value: django}
-  model: questionnaire.choice
-  pk: 9
-- fields: {question: 4, sortid: 2, tags: '', text_de: Pylons, text_en: Pylons, value: pylons}
-  model: questionnaire.choice
-  pk: 10
-- fields: {question: 4, sortid: 3, tags: '', text_de: Turbogears, text_en: Turbogears,
-    value: turbogears}
-  model: questionnaire.choice
-  pk: 11
-- fields: {question: 4, sortid: 4, tags: '', text_de: CherryPy, text_en: CherryPy,
-    value: cherrypy}
-  model: questionnaire.choice
-  pk: 12
-- fields: {question: 5, sortid: 1, tags: '', text_de: Django, text_en: Django, value: django}
-  model: questionnaire.choice
-  pk: 13
-- fields: {question: 5, sortid: 2, tags: '', text_de: Pylons, text_en: Pylons, value: pylons}
-  model: questionnaire.choice
-  pk: 14
-- fields: {question: 5, sortid: 3, tags: '', text_de: Turbogears, text_en: Turbogears,
-    value: turbogears}
-  model: questionnaire.choice
-  pk: 15
-- fields: {question: 5, sortid: 4, tags: '', text_de: CherryPy, text_en: CherryPy,
-    value: cherrypy}
-  model: questionnaire.choice
-  pk: 16
-- fields: {answer: '["yes"]', question: 1, run: 9, subject: 11}
-  model: questionnaire.answer
-  pk: 1
-- fields: {answer: '["weizen"]', question: 2, run: 9, subject: 11}
-  model: questionnaire.answer
-  pk: 2
-- fields: {answer: '["becks", "heineken"]', question: 3, run: 9, subject: 11}
-  model: questionnaire.answer
-  pk: 3
-- fields: {answer: '["django", "pylons", ["rails"]]', question: 4, run: 9, subject: 11}
-  model: questionnaire.answer
-  pk: 4
-- fields: {answer: '["django"]', question: 5, run: 9, subject: 11}
-  model: questionnaire.answer
-  pk: 5
-- fields: {body_de: "Wilkommen zu der Fragebogenbeispielseite!\r\n\r\nWenn Sie einen
-      einfachen Fragebogen sehen wollen, <a href=\"take/1\"> \"tun Sie's!</a>\r\n\r\nSie
-      k\xF6nnen das <a href=\"/admin/\">Admin Interface</a> benutzen, um den Fragebogen
-      zu \xE4ndern.", body_en: "Welcome to the example Questionnaire website!\r\n\r\nIf
-      you wish to take a sample questionnaire, <a href=\"take/1\">go for it!</a>\r\n\r\nUse
-      the <a href=\"/admin/\">Admin Interface</a> to change the questionnaire.", public: true,
-    title_de: '', title_en: Welcome}
-  model: page.page
-  pk: index
