+++ questionnaire/management/commands/make_landing_nonces.py
@@ -3,7 +3,7 @@ from ...models import Landing
 
 
 class Command(BaseCommand):
-    help = "make survey nonces with the specified label"
+    help = "make landing nonces with the specified label"
     args = "<how_many> <label>"
 
     def handle(self, how_many=1, label="no label yet", **options):
