+++ questionnaire/page/views.py
@@ -1,7 +1,6 @@
 # Create your views here.
-from django.shortcuts import render, render_to_response
+from django.shortcuts import render
 from django.conf import settings
-from django.template import RequestContext
 from django import http
 from django.utils import translation
 from .models import Page
@@ -10,12 +9,11 @@ def page(request, page_to_render):
     try:
         p = Page.objects.get(slug=page_to_render, public=True)
     except Page.DoesNotExist:
-        return render(request, "pages/{}.html".format(page_to_render), 
-            { "request" : request,}, 
+        return render(request, "pages/{}.html".format(page_to_render),
+            { "request" : request,},
         )
-    print request.session[translation.LANGUAGE_SESSION_KEY]
-    return render(request, "page.html", 
-            { "request" : request, "page" : p, }, 
+    return render(request, "page.html",
+            { "request" : request, "page" : p, },
         )
 
 def langpage(request, lang, page_to_trans):
