+++ questionnaire/templates/pages/complete.html
@@ -1,7 +1,7 @@
 {% extends "base-questionnaire.html" %}
 {% block questionnaire %}
 <h2>
-    Thanks for completing the survey!  
+    Thanks for completing the questionnaire!  
 </h2>
