+++ questionnaire/templates/pages/generic.html
@@ -7,7 +7,7 @@
 
     <h2>Tell us stuff!</h2>
     <p>We are excited that this is Open Access book and really hope that it will be shared around, read by lots of people all over the world, and used in lots of exciting new ways.  We would love to hear about you and how you are using it - please would you share your story with us using this link:    (it will only take about 3 minutes - and you could just get a free book as well  ….)</p>
-    <p>But comeback again, our survey for {{landing.label}} isn't ready yet.</p>
+    <p>But comeback again, our questionnaire for {{landing.label}} isn't ready yet.</p>
 
 {% endblock %}
