+++ questionnaire/templates/pages/summaries.html
@@ -1,7 +1,7 @@
 {% extends "base-questionnaire.html" %}
 {% block questionnaire %}
 <h1>
-    Survey Results Summary
+    Questionnaire Results Summary
 </h1>
 {% for summary in summaries %}
