+++ questionnaire/templates/questionnaire/complete.en.html
@@ -2,7 +2,7 @@
 {% block questionnaire %}
 
 <h2>
-    Thanks for completing the survey!  
+    Thanks for completing the questionnaire!  
 </h2>
 <div class="question-text">
 {{ landing_object.claim.all.0.rights_holder }}
