+++ questionnaire/templates/questionnaire/questionset.html
@@ -5,7 +5,7 @@
 {% load landings %}
 
 {% block title %}
-        Survey: {{ questionset.heading }}
+        Questionnaire: {{ questionset.heading }}
 {% endblock %}
 
 {% block headextra %}
